refactor program, use custom server and router, restrict access by middleware
